// src/app/core/types/branding.ts
export type Branding = {
  companyName: string
  headOfficeLabel: string
  logoBase64?: string
  logoFileName?: string
  addressLines: string[]
  tel?: string
  fax?: string
  email?: string
  website?: string
  taxId?: string
  defaultFont: string
  defaultPrimaryColor: string
  defaultAccentColor: string
  stampBase64?: string
  stampFileName?: string
}